requirements.txt
fastapi
uvicorn
python-telegram-bot
python-dotenv
google-cloud-firestore
requests
pydantic
pydantic-settings
jinja2
